# FACT_PROXY_OPENAPI.yaml
openapi: 3.0.3
info:
  title: RideHorizon Fact Proxy API
  version: 0.1.0
  description: |
    MVP contract for the RideHorizon iOS app to request a short or long
    ride-safe place fact without storing or sending an OpenAI API key from
    the device.
servers:
  - url: https://ridehorizon.digitalmercenaries.ai
    description: Production Fly.io proxy
  - url: http://127.0.0.1:3000
    description: Local development proxy
tags:
  - name: Health
    description: Proxy health checks.
  - name: Facts
    description: Ride-safe place facts.
  - name: Provisioning
    description: One-time private-alpha device credential provisioning.
  - name: Admin
    description: Operational controls. Not used by the iOS app.
x-deployment:
  provider: Fly.io
  app: ridehorizon-fact-proxy
  organization: dml
  primaryRegion: lhr
  image: ridehorizon-fact-proxy:deployment-01KWFY4N628G4137Y7BMQPN6P9
  machines:
    - id: 080d306c727d98
      process: app
      region: lhr
      state: started
      version: 6
    - id: 8ee01dc77de778
      process: app
      region: lhr
      state: started
      version: 6
  publicIngress:
    sharedIpv4: 66.241.125.198
    dedicatedIpv6: "2a09:8280:1::13b:6469:0"
  deployedSecrets:
    - OPENAI_API_KEY
    - RIDEHORIZON_PROXY_TOKEN
  optionalSecrets:
    - RIDEHORIZON_ADMIN_TOKEN
  runtimeConfiguration:
    RIDEHORIZON_DIAGNOSTICS_ENABLED:
      default: false
      description: Enables verbose proxy diagnostics at startup.
    RIDEHORIZON_PROMPT_OVERRIDES_ENABLED:
      default: false
      description: Load optional prompt overrides from Fly object storage URL.
    RIDEHORIZON_PROMPT_OVERRIDES_OBJECT_URL:
      description: Object-store URL containing optional prompt override JSON.
    RIDEHORIZON_PROMPT_OVERRIDES_REFRESH_SECONDS:
      default: 60
      description: How often proxy reloads prompt overrides from object storage.
    RIDEHORIZON_SHORT_FACT_PROMPT:
      default: built-in short fact prompt
      description: Optional server-side prompt override for shortFacts requests. Never sent by iOS.
    RIDEHORIZON_LONG_FACT_PROMPT:
      default: built-in long fact prompt
      description: Optional server-side prompt override for longFacts requests. Never sent by iOS.
    RIDEHORIZON_DEVICE_BINDING_REQUIRED:
      default: false
      description: When true, requires X-RideHorizon-Device-Id to be in RIDEHORIZON_TRUSTED_DEVICE_IDS.
    RIDEHORIZON_TRUSTED_DEVICE_IDS:
      description: Comma-separated list of allowed X-RideHorizon-Device-Id values when device binding is required.
    RIDEHORIZON_PROMPT_OVERRIDES_HOST_ALLOWLIST:
      description: Comma-separated host allowlist for prompt override JSON fetches. Required when RIDEHORIZON_PROMPT_OVERRIDES_ENABLED=true.
    OPENAI_MODEL:
      default: gpt-4o-mini
      description: OpenAI model selected by the Fly runtime environment.
    RATE_LIMIT_PER_MINUTE:
      default: 30
      description: Request limit for authenticated proxy calls, with user/device identity preference then IP fallback.
x-current-security-model:
  transport: HTTPS through Fly.io public ingress.
  appAuthentication: Revocable per-device bearer credentials, with the shared Fly secret retained temporarily for operator/developer access.
  deviceSecretStorage: iOS Keychain generic-password service RideHorizonProxy.
  serverSecretStorage: Fly.io secrets.
  openAiKeyExposure: Server-side only. Never store the OpenAI API key in the iOS app or repository.
  currentLimitations:
    - The shared RIDEHORIZON_PROXY_TOKEN fallback can call the proxy until rotated or rate-limited.
    - Private-alpha provisioning uses one-time invite codes rather than Apple App Attest or user identity.
  plannedHardening:
    - Per-device registration.
    - Apple App Attest assertion verification.
    - Server-side device approval, revocation, and quota state.
    - Per-user auth before wider non-TestFlight distribution.
x-observability:
  responseHeaders:
    X-Request-Id: Correlates app requests with Fly proxy logs.
  logPrivacy:
    logsTokens: false
    logsOpenAiKey: false
    logsExactPlaceName: false
    logsRiderCoordinates: false
  events:
    - fact_proxy_request
    - fact_request_valid
    - fact_request_success
    - fact_request_rejected
    - proxy_auth_failed
    - proxy_auth_misconfigured
    - rate_limit_exceeded
    - openai_response
    - openai_upstream_error
    - openai_request_failed
    - diagnostics_updated
    - prompt_overrides_loaded
    - prompt_overrides_load_failed
    - prompt_overrides_load_skipped
  diagnostics:
    baselineRequestLogsAlwaysEnabled: true
    verboseDiagnosticsDefault: false
    startupEnvironmentVariable: RIDEHORIZON_DIAGNOSTICS_ENABLED
    adminEndpoint: /admin/diagnostics
    adminEndpointRequiresSecret: RIDEHORIZON_ADMIN_TOKEN
    adminEndpointHiddenWhenSecretUnset: true
    runtimeChangesAreProcessLocal: true
paths:
  /health:
    get:
      tags:
        - Health
      summary: Check proxy health
      description: Public health check. Does not require bearer authentication.
      operationId: getHealth
      security: []
      responses:
        "200":
          description: Proxy is running.
          content:
            text/plain:
              schema:
                type: string
                enum:
                  - ok
  /v1/fact:
    post:
      tags:
        - Facts
      summary: Generate one ride-safe place fact
      description: |
        Validates a place request, calls the server-side LLM provider, sanitizes
        the generated text, and returns one bounded fact. The iOS app sends
        only `factMode`, boundary/place fields, and place hierarchy. It never
        sends arbitrary model messages, OpenAI configuration, ElevenLabs
        configuration, or coordinates. The proxy owns prompt selection for
        `shortFacts` and `longFacts`. Optional rider context text, home region
        preferences, and fact-interest categories are treated as bounded user
        preference data, not as system prompt text. The iOS app must fall back
        to the base place announcement
        for any non-200 response, invalid response body, rejected fact, missing
        token, or request timeout.
      operationId: createPlaceFact
      security:
        - RideHorizonProxyToken: []
      parameters:
        - in: header
          name: X-RideHorizon-User-Id
          description: Optional rider/user identifier for server-side prompt override lookup.
          required: false
          schema:
            type: string
            minLength: 4
            maxLength: 64
        - in: header
          name: X-RideHorizon-Device-Id
          description: |
            Optional approved-device identifier. Required when
            RIDEHORIZON_DEVICE_BINDING_REQUIRED=true.
          required: false
          schema:
            type: string
            minLength: 4
            maxLength: 64
      x-ios-keychain-service: RideHorizonProxy
      x-ios-device-id-keychain-service: RideHorizonDeviceId
      x-ios-timeout-seconds: 15
      x-ios-fallback: Speak the base place announcement without a fact.
      x-openai-api-key-location: Server-side only. Never store the OpenAI API key in the iOS app or repository.
      x-speech-safety-rules:
        - Short Facts may use up to 5 concise sentences.
        - Long Facts may use up to 8 concise sentences.
        - Short Facts prioritises geographic/cultural context over rider-coaching details.
        - Factual and neutral.
        - Useful as ambient place context.
        - Short enough to keep the total spoken announcement ride-safe.
        - No questions.
        - No invitations, route advice, speed advice, or riding instructions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FactRequest"
            examples:
              town:
                summary: UK town request
                value:
                  boundary: town
                  placeName: Stroud
                  factMode: shortFacts
                  countryContext: United Kingdom
                  placeHierarchy:
                    street: null
                    town: Stroud
                    county: Gloucestershire
                    region: England
                    country: United Kingdom
                  riderContext:
                    homeCountry: United Kingdom
                    homeRegion: West Midlands
                    familiarRegions:
                      - England
                      - Cotswolds
                    factInterestCategories:
                      - geographyBasics
                      - locationFacts
                      - pointsOfInterest
                      - history
                    customFactInstructions: engineering and old roads
      responses:
        "200":
          description: One sanitized bounded fact.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FactResponse"
              examples:
                fact:
                  value:
                    fact: Known for its wool trade.
        "400":
          description: Invalid JSON, missing required field, invalid boundary, invalid factMode, invalid placeName, or invalid placeHierarchy.
        "401":
          description: Missing or wrong proxy bearer token.
        "415":
          description: Unsupported content type. Proxy only accepts application/json.
        "500":
          description: Proxy is misconfigured, including missing server-side proxy token.
        "502":
          description: OpenAI returned an error or unusable response.
  /v1/speech:
    post:
      tags:
        - Speech
      summary: Generate ride-safe speech audio through ElevenLabs
      description: |
        Validates bounded rider-facing text, calls ElevenLabs server-side, and
        returns MP3 audio. The iOS app sends only the text to speak. It never
        sends an ElevenLabs API key, voice id, model id, output format, prompt
        text, or coordinates. The proxy owns all ElevenLabs configuration.
      operationId: createSpeech
      security:
        - RideHorizonProxyToken: []
      parameters:
        - in: header
          name: X-RideHorizon-Device-Id
          description: |
            Optional approved-device identifier. Required when
            RIDEHORIZON_DEVICE_BINDING_REQUIRED=true.
          required: false
          schema:
            type: string
            minLength: 4
            maxLength: 64
      x-ios-keychain-service: RideHorizonProxy
      x-ios-device-id-keychain-service: RideHorizonDeviceId
      x-ios-timeout-seconds: 15
      x-ios-fallback: Fall back to local Apple AVSpeechSynthesizer speech.
      x-elevenlabs-api-key-location: Server-side only. Never store the ElevenLabs API key in the iOS app or repository.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeechRequest"
            examples:
              phrase:
                value:
                  text: Stroud was known for its wool trade.
      responses:
        "200":
          description: MP3 audio for the submitted speech text.
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid JSON, missing text, or overlong text.
        "401":
          description: Missing or wrong proxy bearer token.
        "415":
          description: Unsupported content type. Proxy only accepts application/json.
        "500":
          description: Proxy is misconfigured, including missing server-side proxy token.
        "502":
          description: ElevenLabs returned an error or unusable response. Raw provider/account details are never returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeechErrorResponse"
  /v1/provision:
    post:
      tags:
        - Provisioning
      summary: Redeem a one-time invite for a device credential
      operationId: provisionDeviceCredential
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvisionRequest"
      responses:
        "201":
          description: Credential issued once. The plaintext credential is never returned again.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisionResponse"
        "400":
          description: Invalid JSON.
        "401":
          description: Invite is invalid, expired, or already consumed.
        "429":
          description: Provisioning rate limit exceeded.
  /admin/v1/invites:
    post:
      tags:
        - Admin
      summary: Create a one-time private-alpha invite
      operationId: createPrivateAlphaInvite
      security:
        - RideHorizonAdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviteRequest"
      responses:
        "201":
          description: Invite created. The plaintext invite code is returned once.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteResponse"
        "401":
          description: Missing or wrong administrative bearer credential.
        "404":
          description: Administrative authentication is not configured.
  /admin/v1/credentials:
    get:
      tags:
        - Admin
      summary: List issued credential metadata
      operationId: listDeviceCredentials
      security:
        - RideHorizonAdminToken: []
      responses:
        "200":
          description: Credential metadata without plaintext credentials or hashes.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CredentialSummary"
        "401":
          description: Missing or wrong administrative bearer credential.
  /admin/v1/credentials/{credentialId}:
    delete:
      tags:
        - Admin
      summary: Revoke a device credential
      operationId: revokeDeviceCredential
      security:
        - RideHorizonAdminToken: []
      parameters:
        - in: path
          name: credentialId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Credential revoked.
        "401":
          description: Missing or wrong administrative bearer credential.
        "404":
          description: Credential was not found.
  /admin/diagnostics:
    get:
      tags:
        - Admin
      summary: Read verbose diagnostics state
      description: |
        Operational endpoint. Not used by the iOS app. Requires
        `RIDEHORIZON_ADMIN_TOKEN`. If that token is not configured, this endpoint
        returns 404.
      operationId: getDiagnostics
      security:
        - RideHorizonAdminToken: []
      responses:
        "200":
          description: Current diagnostics state for this proxy process.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticsResponse"
        "401":
          description: Missing or wrong admin bearer token.
        "404":
          description: Admin diagnostics endpoint is disabled because RIDEHORIZON_ADMIN_TOKEN is not configured.
    put:
      tags:
        - Admin
      summary: Update verbose diagnostics state
      description: |
        Operational endpoint. Not used by the iOS app. Runtime changes are
        process-local; on multi-machine deployments, prefer
        `RIDEHORIZON_DIAGNOSTICS_ENABLED` for a consistent fleet-wide setting.
      operationId: updateDiagnostics
      security:
        - RideHorizonAdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagnosticsRequest"
      responses:
        "200":
          description: Updated diagnostics state for this proxy process.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticsResponse"
        "401":
          description: Missing or wrong admin bearer token.
        "404":
          description: Admin diagnostics endpoint is disabled because RIDEHORIZON_ADMIN_TOKEN is not configured.
components:
  securitySchemes:
    RideHorizonProxyToken:
      type: http
      scheme: bearer
      bearerFormat: RIDEHORIZON_PROXY_TOKEN
      description: |
        Per-device credential issued by `/v1/provision`. The iOS app reads it
        from the Keychain generic-password item with service `RideHorizonProxy`.
        The shared Fly secret remains a temporary operator/developer fallback.
    RideHorizonAdminToken:
      type: http
      scheme: bearer
      bearerFormat: RIDEHORIZON_ADMIN_TOKEN
      description: |
        Separate operational token when configured. During private alpha only,
        the server-side shared proxy token is the administrative fallback. Neither
        credential is shipped in the iOS app.
  schemas:
    CreateInviteRequest:
      type: object
      additionalProperties: false
      properties:
        label:
          type: string
          maxLength: 80
          description: Human-readable tester label. Must not contain secret material.
    InviteResponse:
      type: object
      additionalProperties: false
      required:
        - inviteId
        - inviteCode
        - expiresAt
      properties:
        inviteId:
          type: string
          format: uuid
        inviteCode:
          type: string
          description: One-time secret returned only in this response.
        expiresAt:
          type: string
          format: date-time
    ProvisionRequest:
      type: object
      additionalProperties: false
      required:
        - inviteCode
        - deviceId
      properties:
        inviteCode:
          type: string
          minLength: 4
          maxLength: 128
        deviceId:
          type: string
          minLength: 4
          maxLength: 64
          pattern: "^[a-z0-9._:-]+$"
    ProvisionResponse:
      type: object
      additionalProperties: false
      required:
        - credentialId
        - credential
        - expiresAt
      properties:
        credentialId:
          type: string
          format: uuid
        credential:
          type: string
          description: Device-bound secret returned only in this response and stored in iOS Keychain.
        expiresAt:
          type: string
          format: date-time
    CredentialSummary:
      type: object
      additionalProperties: false
      required:
        - credentialId
        - deviceId
        - createdAt
        - expiresAt
      properties:
        credentialId:
          type: string
          format: uuid
        deviceId:
          type: string
        label:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        revokedAt:
          type: string
          format: date-time
          nullable: true
    Boundary:
      type: string
      description: Boundary type that triggered the announcement.
      enum:
        - country
        - nation
        - county
        - town
        - street
    FactRequest:
      type: object
      additionalProperties: false
      required:
        - boundary
        - placeName
        - factMode
        - placeHierarchy
      properties:
        boundary:
          $ref: "#/components/schemas/Boundary"
        placeName:
          type: string
          minLength: 1
          maxLength: 96
          pattern: "^[A-Za-z0-9 .,'’&()\\-]+$"
          description: Place to generate a fact about.
          example: Stroud
        factMode:
          type: string
          description: |
            Requested fact depth. The proxy validates this field and selects
            the server-side prompt. Unknown values return 400.
          enum:
            - shortFacts
            - longFacts
          example: shortFacts
        countryContext:
          type: string
          nullable: true
          minLength: 1
          maxLength: 64
          pattern: "^[A-Za-z .,'’()\\-]+$"
          description: Optional country name used to disambiguate reused place names.
          example: United Kingdom
        placeHierarchy:
          $ref: "#/components/schemas/PlaceHierarchy"
        riderContext:
          $ref: "#/components/schemas/RiderContext"
    PlaceHierarchy:
      type: object
      additionalProperties: false
      description: |
        Current place hierarchy from iOS reverse geocoding. Values may be null
        when unavailable or invalid. Coordinates are intentionally not part of
        this contract.
      properties:
        street:
          $ref: "#/components/schemas/OptionalPlaceName"
        town:
          $ref: "#/components/schemas/OptionalPlaceName"
        county:
          $ref: "#/components/schemas/OptionalPlaceName"
        region:
          $ref: "#/components/schemas/OptionalPlaceName"
        country:
          allOf:
            - $ref: "#/components/schemas/OptionalCountryName"
    OptionalPlaceName:
      type: string
      nullable: true
      minLength: 1
      maxLength: 96
      pattern: "^[A-Za-z0-9 .,'’&()\\-]+$"
      example: Nailsworth
    OptionalCountryName:
      type: string
      nullable: true
      minLength: 1
      maxLength: 64
      pattern: "^[A-Za-z .,'’()\\-]+$"
      example: United Kingdom
    RiderContext:
      type: object
      additionalProperties: false
      description: Optional rider-context hints sent to the server for fact generation.
      properties:
        homeCountry:
          $ref: "#/components/schemas/OptionalCountryName"
          description: Optional country of origin to avoid obvious admin-level context.
        homeRegion:
          $ref: "#/components/schemas/OptionalPlaceName"
          description: Optional region known as familiar to the rider.
        familiarRegions:
          type: array
          maxItems: 12
          items:
            $ref: "#/components/schemas/OptionalPlaceName"
          description: Optional familiar regions to avoid repeating generic explanations.
        factInterestCategories:
          type: array
          maxItems: 7
          items:
            type: string
            enum:
              - localRidingHints
              - safetyAdvice
              - geographyBasics
              - locationFacts
              - pointsOfInterest
              - history
              - culture
              - landmarks
          description: Optional rider-selected fact categories.
        customFactInstructions:
          type: string
          nullable: true
          minLength: 1
          maxLength: 240
          pattern: "^[A-Za-z0-9 .,'’&()\\-]+$"
          description: |
            Optional rider preference hint for fact focus, such as "old roads,
            engineering, local industry". This is untrusted user preference
            data, not system prompt text. The proxy validates and wraps it in
            fixed safety instructions.
      example:
        homeCountry: United Kingdom
        homeRegion: West Midlands
        familiarRegions:
          - England
          - Cotswolds
        factInterestCategories:
          - geographyBasics
          - locationFacts
          - pointsOfInterest
          - history
        customFactInstructions: engineering and old roads
    SpeechRequest:
      type: object
      additionalProperties: false
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 1400
          description: Bounded rider-facing text to synthesize as speech.
          example: Stroud was known for its wool trade.
    SpeechErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
        - code
      properties:
        error:
          type: string
          enum:
            - Premium voice is temporarily unavailable.
          description: Rider-safe message that never contains provider account, credit, quota, authentication, or billing details.
        code:
          type: string
          enum:
            - RH-TTS-01
            - RH-TTS-02
            - RH-TTS-03
            - RH-TTS-04
          description: Opaque RideHorizon diagnostic code for developer diagnostics.
    FactResponse:
      type: object
      additionalProperties: false
      required:
        - fact
      properties:
        fact:
          type: string
          minLength: 1
          maxLength: 1500
      description: |
        Sanitized ride-safe fact. shortFacts responses are bounded to 1100
        characters; longFacts responses are bounded to 1500 characters.
          example: Known for its wool trade.
    DiagnosticsRequest:
      type: object
      additionalProperties: false
      required:
        - enabled
      properties:
        enabled:
          type: boolean
          description: Whether verbose diagnostics should be enabled for this proxy process.
    DiagnosticsResponse:
      type: object
      additionalProperties: false
      required:
        - enabled
      properties:
        enabled:
          type: boolean
          description: Whether verbose diagnostics are enabled for this proxy process.